Refactor DGPU support code into drivers/system76/dgpu
