+++ src/drivers/system76/dgpu/Makefile.inc
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: GPL-2.0-only
+
+ramstage-$(CONFIG_DRIVERS_SYSTEM76_DGPU) += ramstage.c
