+++ src/mainboard/system76/addw2/bootblock.c
@@ -1,29 +1,11 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
 
 #include <bootblock_common.h>
-#include <console/console.h>
-#include <delay.h>
 #include <gpio.h>
 #include "gpio.h"
+#include <drivers/system76/dgpu/bootblock.c>
 
-static void dgpu_power_enable(int onoff) {
-	printk(BIOS_DEBUG, "system76: DGPU power %d\n", onoff);
-	if (onoff) {
-		gpio_set(DGPU_RST_N, 0);
-		mdelay(4);
-		gpio_set(DGPU_PWR_EN, 1);
-		mdelay(4);
-		gpio_set(DGPU_RST_N, 1);
-	} else {
-		gpio_set(DGPU_RST_N, 0);
-		mdelay(4);
-		gpio_set(DGPU_PWR_EN, 0);
-	}
-	mdelay(50);
-}
-
-void bootblock_mainboard_init(void)
-{
+void bootblock_mainboard_init(void) {
 	gpio_configure_pads(early_gpio_table, ARRAY_SIZE(early_gpio_table));
 	dgpu_power_enable(1);
 }
