+++ src/mainboard/system76/addw2/romstage.c
@@ -78,8 +78,7 @@ static const struct cnl_mb_cfg memcfg = {
 	.ect = 0,
 };
 
-void mainboard_memory_init_params(FSPM_UPD *memupd)
-{
+void mainboard_memory_init_params(FSPM_UPD *memupd) {
 	// Allow memory clocks higher than 2933 MHz
 	memupd->FspmConfig.SaOcSupport = 1;
 	// Set primary display to internal graphics
