+++ src/mainboard/system76/gaze14/Kconfig
@@ -4,6 +4,7 @@ config BOARD_SPECIFIC_OPTIONS
 	def_bool y
 	select BOARD_ROMSIZE_KB_16384
 	select DRIVERS_I2C_HID
+	select DRIVERS_SYSTEM76_DGPU
 	select EC_SYSTEM76_EC
 	select HAVE_ACPI_RESUME
 	select HAVE_ACPI_TABLES
