+++ src/mainboard/system76/gaze14/acpi/mainboard.asl
@@ -1,5 +1,8 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
 
+#include "../gpio.h"
+#include <drivers/system76/dgpu/acpi/dgpu.asl>
+
 #define EC_GPE_SCI 0x03 /* GPP_K3 */
 #define EC_GPE_SWI 0x06 /* GPP_K6 */
 #define EC_COLOR_KEYBOARD 1
@@ -9,8 +12,6 @@ Scope (\_SB) {
 	#include "sleep.asl"
 	Scope (PCI0) {
 		#include "backlight.asl"
-		#include "pegp.asl"
-		#include "dgpu.asl"
 	}
 }
