+++ src/mainboard/system76/oryp6/Kconfig
@@ -5,6 +5,7 @@ config BOARD_SPECIFIC_OPTIONS
 	select BOARD_ROMSIZE_KB_16384
 	select DRIVERS_I2C_HID
 	select DRIVERS_I2C_TAS5825M
+	select DRIVERS_SYSTEM76_DGPU
 	select EC_SYSTEM76_EC
 	select HAVE_ACPI_RESUME
 	select HAVE_ACPI_TABLES
